# frontend/src/assets/openApi/documentation.yaml
openapi: "3.1.0"
info:
  title: "SAAS API Documentation"
  version: "1.0.0"
  description: "SAAS API Documentation"
  contact:
    name: "SAAS API Documentation"
    email: "contact@limpidcrypto.com"
servers:
  - url: https://xrp.limpidcrypto.com/api/v1/Test
    description: "Testnet API"
  - url: https://xrp.limpidcrypto.com/api/v1/Dev
    description: "Devnet API"
  - url: https://xrp.limpidcrypto.com/api/v1/XahauTest
    description: "Xahau Testnet API"
  - url: https://xrp.limpidcrypto.com/api/v1/SidechainTest
    description: "Sidechain Testnet API"
tags:
  - name: wallet
    description: Manage XRP wallets
schemes:
  - https
paths:
  /wallet/faucet:
    get:
      summary: Generates a new wallet and funds it
      description: Generates a new wallet and funds it on the selected network
      operationId: generate_faucet_wallet
      tags:
        - wallet
      responses:
        "200":
          description: successfully generated a new faucet wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Faucet"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
    Account:
      type: object
      required:
        - classic_address
      properties:
        classic_address:
          type: string
        address:
          type: string
        xaddress:
          type: string
        secret:
          type: string
    Faucet:
      type: object
      required:
        - account
        - amount
      properties:
        account:
          $ref: "#/components/schemas/Account"
        amount:
          type: integer
          format: int32
          minimum: 0
        balance:
          type: integer
          format: int32
          minimum: 0
        trace:
          $ref: "#/components/schemas/Trace"
    Trace:
      type: object
      properties:
        hash:
          type: string
          format: hash
        code:
          type: string